chore: standarlize pearson example
